+++ demo/HuggingFace/GPT2/GPT2ModelConfig.py
@@ -1,11 +1,12 @@
 #
-# Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 1993-2022 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -68,6 +69,9 @@ class GPT2Metadata(_GPT2Metadata, MetadataArgparseInteropMixin):
         base_metadata = GPT2Metadata.from_args(args)
         return base_metadata._replace(precision=Precision(fp16=args.fp16))
 
+    @staticmethod
+    def add_benchmarking_args(parser: argparse.ArgumentParser) -> None:
+        pass
 
 class GPT2ModelTRTConfig(NNConfig):
     VOCAB_SIZE = 50257  # Vocabulary size of the GPT-2 model