+++ plugin/common/kernels/maskRCNNKernels.h
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+ * SPDX-FileCopyrightText: Copyright (c) 1993-2022 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -271,7 +272,8 @@ struct xy_t
     }
 };
 // PYRAMID ROIALIGN
-cudaError_t roiAlign(cudaStream_t stream, int batchSize, int featureCount, int roiCount, float firstThreshold,
+cudaError_t roiAlign(cudaStream_t stream, int batchSize, xy_t imageSize, int featureCount, int roiCount,
+    float firstThreshold, int transformCoords, bool absCoords, bool swapCoords, bool plusOneCoords, int samplingRatio,
 
     const void* rois, const void* const layers[], const xy_t* layerDims,