+++ python/src/infer/pyPlugin.cpp
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+ * SPDX-FileCopyrightText: Copyright (c) 1993-2022 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,6 +18,7 @@
 // This file contains all bindings related to plugins.
 #include "ForwardDeclarations.h"
 #include "infer/pyPluginDoc.h"
+#include "utils.h"
 #include <cuda_runtime_api.h>
 #include <pybind11/stl.h>
 
@@ -217,7 +219,9 @@ void bindPlugin(py::module& m)
         .def("__len__", [](PluginFieldCollection& self) { return self.nbFields; })
         .def("__getitem__", [](PluginFieldCollection& self, int index) {
             if (index >= self.nbFields)
-                throw py::index_error();
+            {
+                utils::throwPyIndexError(); // See definition of throwPyIndexError() for details
+            }
             return self.fields[index];
         });